Switch to pinned dependency versions.
This is in preparation for a bulk update - Github notified me that there are a ton of outdated deps here (which is true), so I'm using this as an opportunity to practice handling stale deps.

Part of the motivation here is for `cargo upgrade` (from `cargo-edit`) to be able to do something reasonable.